requirements.txt
pydantic 
pydantic-settings
python-dotenv
langgraph
langchain 
langsmith
langchain-text-splitters 
langchain-community 
bs4
langchain-core
langchain-google-genai
langchain-tavily
scrapegraph-py
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
asyncpg
psycopg[binary]
psycopg2-binary
httpx
celery[redis]
redis